refactor: move donors routes to new folder structure
