+++ src/API/REST/V3/Routes/Donors/RegisterDonorRoutes.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Give\Donors\Routes;
+namespace Give\API\REST\V3\Routes\Donors;
 
+use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorRoute;
 use Give\Donors\Controllers\DonorRequestController;
 use Give\Donors\ValueObjects\DonorAnonymousMode;
-use Give\Donors\ValueObjects\DonorRoute;
 use WP_Error;
 use WP_REST_Request;
 use WP_REST_Server;
