+++ tests/Unit/API/REST/V3/Routes/Donations/GetDonationsRouteTest.php
@@ -5,12 +5,12 @@ namespace Unit\API\REST\V3\Routes\Donations;
 use Exception;
 use Give\API\REST\V3\Routes\Donations\RegisterDonationRoutes;
 use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
+use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorRoute;
 use Give\Campaigns\Models\Campaign;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationMetaKeys;
 use Give\Donations\ValueObjects\DonationMode;
 use Give\Donations\ValueObjects\DonationStatus;
-use Give\Donors\ValueObjects\DonorRoute;
 use Give\Framework\Support\ValueObjects\Money;
 use Give\Helpers\Hooks;
 use Give\Tests\RestApiTestCase;
