+++ tests/Unit/API/REST/V3/Routes/Donors/GetDonorsRouteTest.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Unit\Donors\Routes;
+namespace Unit\API\REST\V3\Routes\Donors;
 
 use Exception;
+use Give\API\REST\V3\Routes\Donors\RegisterDonorRoutes;
+use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorRoute;
 use Give\Campaigns\Models\Campaign;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationMetaKeys;
 use Give\Donations\ValueObjects\DonationMode;
 use Give\Donations\ValueObjects\DonationStatus;
 use Give\Donors\Models\Donor;
-use Give\Donors\Routes\RegisterDonorRoutes;
-use Give\Donors\ValueObjects\DonorRoute;
 use Give\Framework\Database\DB;
 use Give\Framework\Support\ValueObjects\Money;
 use Give\Helpers\Hooks;
